Refactor repository structure to use interfaces
This commit refactors the repository interfaces and implementations to adopt a new structure that includes `Interfaces` for repositories and services. Several controller files have been updated to use the new `Interfaces` namespace, promoting a more interface-driven design.

New repository interfaces have been added or modified to define method signatures for various operations, ensuring clear contracts for implementations. The corresponding repository implementations have been updated to align with these interfaces.

Additionally, the `IStorageService` interface has been introduced to manage blob storage operations, enhancing the API for file handling.

These changes improve code maintainability and testability by facilitating easier mocking and dependency injection in unit tests.
