+++ src/backend/FleetAssistant.WebApi/Program.cs
@@ -4,7 +4,9 @@ using FleetAssistant.WebApi.Data;
 using FleetAssistant.WebApi.HealthChecks;
 using FleetAssistant.WebApi.Options;
 using FleetAssistant.WebApi.Repositories;
+using FleetAssistant.WebApi.Repositories.Interfaces;
 using FleetAssistant.WebApi.Services;
+using FleetAssistant.WebApi.Services.Interfaces;
 using Microsoft.AspNetCore.Diagnostics.HealthChecks;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.OpenApi.Models;
