+++ src/backend/FleetAssistant.WebApi/Repositories/FuelLogRepository.cs
@@ -1,5 +1,6 @@
 using FleetAssistant.Shared.Models;
 using FleetAssistant.WebApi.Data;
+using FleetAssistant.WebApi.Repositories.Interfaces;
 using Microsoft.EntityFrameworkCore;
 
 namespace FleetAssistant.WebApi.Repositories;
