+++ src/backend/FleetAssistant.WebApi/Repositories/Interfaces/IMaintenanceRepository.cs
@@ -1,6 +1,6 @@
 using FleetAssistant.Shared.Models;
 
-namespace FleetAssistant.WebApi.Repositories;
+namespace FleetAssistant.WebApi.Repositories.Interfaces;
 
 /// <summary>
 /// Repository interface for maintenance record operations
