+++ src/backend/FleetAssistant.WebApi/Repositories/Interfaces/IVehicleRepository.cs
@@ -1,6 +1,6 @@
 using FleetAssistant.Shared.Models;
 
-namespace FleetAssistant.WebApi.Repositories;
+namespace FleetAssistant.WebApi.Repositories.Interfaces;
 
 /// <summary>
 /// Repository interface for Vehicle entities with specialized operations
